requirements.txt
pandas==1.2.1
plotly==4.14.1
SQLAlchemy==1.3.13
PyMySQL==1.0.2
dash==2.2.0
PyGithub==1.55
requests==2.25.1